Merge of [293] from branches/mac:
In [128], Make.py was changed to generate a Makefile which links against
the dynamic library of Python. This looks like a mistake, since the
distribution than depends on libpythonX.X.so.
I changed it back, so that it still uses syconfig to extract the correct
path of the static library. The problem now appears to be that not all
distributions include a static library...


git-svn-id: http://svn.pyinstaller.org/trunk@297 8dd32b29-ccff-0310-8a9a-9233e24343b1