Introduce printf()-style error macros in bootloader for Windows.
From: Anton Gyllenberg <anton@iki.fi>

The output macros -- VS(), FATALERROR() and OTHERERROR(), had no output
formatting options. This lead to unnatural constructs like
	VS("Loading dll: ");
	VS(dllpath);
	VS("\n");

Worse, for a Windows windowed application, the above example would
generate three dialog boxes of which the last one blank.

Therefore, make the functions behave like and have the same signature
as printf(). They should be backwards compatible as the old invocations
had just one argument which for some configurations was used as a format
string to printf.

I could not find a way to use preprocessor macros with a variable
amount of arguments on Microsoft Visual Studio 2003, so I opted to
make real functions for the message boxes on Windows.


git-svn-id: http://svn.pyinstaller.org/trunk@329 8dd32b29-ccff-0310-8a9a-9233e24343b1